Refactor expense handling to use decimal types for financial calculations
- Updated `total` and `quantity` fields in the database schema and related DTOs to use `numeric(12,2)` and `decimal` types for improved precision in financial calculations.
- Adjusted logic in `CreateExpenseHeaderHandler` and `UpdateExpenseHeaderHandler` to calculate totals based on the new decimal types, ensuring accurate expense summaries.
- Enhanced validation to prevent bank payments from exceeding total expenses, improving business logic integrity.

+++ SenorArroz.Application/Features/ExpenseHeaders/DTOs/ExpenseDetailDto.cs
@@ -8,9 +8,9 @@ public class ExpenseDetailDto
     public string ExpenseName { get; set; } = string.Empty;
     public string ExpenseCategoryName { get; set; } = string.Empty;
     public string ExpenseUnit { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public int Amount { get; set; }
-    public int? Total { get; set; }
+    public decimal? Total { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
